JobMatcherBot/requirements.txt
python-telegram-bot==20.8
beautifulsoup4==4.12.3
requests==2.31.0
pdfminer.six==20221105
nltk==3.8.1
openai==1.30.1
python-docx==1.1.0